Rename component, add mobile styles
